se creo combineReducer y filtrado con ordenamiento
