+++ src/state/ducks/detail/reducer.js
@@ -1,0 +1,20 @@
+  const initialState = {
+    
+  }
+
+
+
+
+   function detailReducer(state = initialState, action){
+      switch(action.type){
+        case "GET_DETAIL":
+      return {
+        ...action.payload,
+      };
+
+      default: return state;
+      }
+   }
+
+   export default detailReducer;
+
