+++ src/state/ducks/professionals/actions.js
@@ -83,20 +83,7 @@ export function orderByName(payload) {
     }
 }
 
-export function getDetail(id) {
-    return async function (dispatch) {
-        try {
-            var detalle = await axios.get('/home/professionals/' + id)
 
-            return dispatch({
-                type: 'GET_DETAIL',
-                payload: detalle.data
-            })
-        } catch (error) {
-            console.log(error)
-        }
-    }
-}
 
 export function getOrderReputation(payload) {
     return {
