+++ src/state/ducks/professionals/reducer.js
@@ -42,18 +42,7 @@ function rootReducer(state = initialState, action) {
 return{...state,
   professionalsFiltered:state.allProfessionals
 }
-    // case "GET_PRFESSIONS_OF_PROFESSIONALS":
-    //   let allProfessionalsProfessions = state.allProfessionals;
-    //   const filtrados = action.payload === "all"
-    //       ? allProfessionalsProfessions
-    //       : allProfessionalsProfessions.filter((el) =>{
-    //         el.professions.includes(action.payload)
-    //       }
-    //           )
-    //   return {
-    //     ...state,
-    //     professionals: filtrados
-    //   };
+  
 
     case "GET_ALL_PROFESSIONS":
       return {
@@ -64,7 +53,7 @@ return{...state,
     case "ORDER_BY_NAME":
       let sortOrder =
         action.payload === "asc"
-          ? state.allProfessionals.sort(function (a, b) {
+          ? state.professionalsFiltered.sort(function (a, b) {
               if (a.firstName > b.firstName) {
                 return 1;
               }
@@ -73,7 +62,7 @@ return{...state,
               }
               return 0;
             })
-          : state.allProfessionals.sort(function (a, b) {
+          : state.professionalsFiltered.sort(function (a, b) {
               if (a.firstName > b.firstName) {
                 return -1;
               }
@@ -86,7 +75,7 @@ return{...state,
         ...state,
         professionals: sortOrder,
       };
-
+     // se saco detail
     case "GET_DETAIL":
       return {
         ...state,
@@ -96,7 +85,7 @@ return{...state,
     case "ORDER_REPUTATION":
       let sortOrder2 =
         action.payload === "asc"
-          ? state.allProfessionals.sort(function (a, b) {
+          ? state.professionalsFiltered.sort(function (a, b) {
               if (a.review > b.review) {
                 return 1;
               }
@@ -105,7 +94,7 @@ return{...state,
               }
               return 0;
             })
-          : state.allProfessionals.sort(function (a, b) {
+          : state.professionalsFiltered.sort(function (a, b) {
               if (a.review > b.review) {
                 return -1;
               }
