+++ src/state/store.js
@@ -2,5 +2,6 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import rootReducer from "./ducks/professionals/reducer";
+import reducers from "../../src/state/ducks/index"
 
-export const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)))
+export const store = createStore(reducers, composeWithDevTools(applyMiddleware(thunk)))
