+++ src/views/components/CardsLists.jsx
@@ -12,8 +12,9 @@ import FormProfession from './Formularios/FormProfession';
 
 
 function CardsList() {
-  const profesionales = useSelector((state) => (state.professionalsFiltered))
-     const professions = useSelector((state)=>(state.professions))
+  const profesionales = useSelector((state) => (state.professionals.professionalsFiltered))
+  console.log(profesionales)
+     const professions = useSelector((state)=>(state.professionals.professions))
 
   
   const dispatch = useDispatch();
