+++ src/views/components/Detail.jsx
@@ -5,7 +5,7 @@ import { useParams } from "react-router-dom"
 import { useEffect } from "react"
 
 //Redux
-import { getDetail } from "../../state/ducks/professionals/actions"
+import { getDetail } from "../../state/ducks/detail/actions"
 
 //CSS Styles
 import "./Detail.css"
@@ -20,6 +20,7 @@ export default function Detail () {
     const { id } = useParams()
 
     const detalle = useSelector((state) => state.detail)
+    console.log(detalle)
 
     const dispatch = useDispatch()
