+++ src/views/pages/Home.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import SearchBar from "../components/SearchBar";
-import { getAllProfessions } from "../../state/ducks/professionals/actions";
+import { getAllProfessionals, getAllProfessions } from "../../state/ducks/professionals/actions";
 
 import CardsList from "../components/CardsLists";
 
@@ -13,6 +13,7 @@ export default function Home() {
 
   useEffect(() => {
     dispatch(getAllProfessions());
+    dispatch(getAllProfessionals())
   }, [dispatch]);
 
   return <CardsList />;
